+++ packages/button/utils/index.ts
@@ -1,1 +1,0 @@
-export { renderRipple, type RippleParams } from "./ripple";