+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aristobyte-ui/docs
 
+## 1.0.22
+
+### Patch Changes
+
+- feat: use utilities from utils package
+
+- Updated dependencies []:
+  - @aristobyte-ui/message-box@1.0.22
+  - @aristobyte-ui/dropdown@1.0.22
+  - @aristobyte-ui/spinner@1.0.22
+  - @aristobyte-ui/anchor@1.0.22
+  - @aristobyte-ui/button@1.0.22
+  - @aristobyte-ui/switch@1.0.22
+  - @aristobyte-ui/radio@1.0.22
+
 ## 1.0.20
 
 ### Patch Changes