+++ packages/dropdown/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aristobyte-ui/dropdown
 
+## 1.0.22
+
+### Patch Changes
+
+- feat: use utilities from utils package
+
+- Updated dependencies []:
+  - @aristobyte-ui/button@1.0.22
+  - @aristobyte-ui/utils@1.0.22
+
 ## 1.0.21
 
 ### Patch Changes