+++ packages/switch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aristobyte-ui/switch
 
+## 1.0.22
+
+### Patch Changes
+
+- feat: use utilities from utils package
+
 ## 1.0.21
 
 ### Patch Changes